refactor: update object model includes in tracker models
